refactor: cargar shaders desde archivo, arreglar RAII, eliminar dead code
- Mover shader inline (108 líneas) de renderer.cpp a shaders/eq_columns.frag
- Corregir default.vert para que coincida con el vertex shader por defecto de GStreamer
- GLRenderer::configure() ahora usa vertexSrc y fragmentSrc en vez de código inline
- Validar width_/height_ > 0 en setTextureSize() y updateAmplitudes()
- Nombres de uniforms como std::array<std::string_view> en vez de C-array
- Pipeline: guardar y limpiar source IDs de g_unix_signal_add en destructor
- Pipeline: eliminar miembros muertos (interrupted_, hasError_, spectrumFrameCount_)
- Pipeline: eliminar debug output cada 30 frames
- Pipeline: añadir manejo de errores GL shader (GST_MESSAGE_ELEMENT)
- GUI: RAII para GtkApplication (g_object_unref en destructor, no en run())
- GUI: move constructor/assignment explícitamente = delete
- SpectrumAnalyzer: eliminar static_cast<float> redundante en g_value_get_float

+++ src/spectrum.cpp
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: shurtado <samuel@hurtadom.dev>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/06/11 15:07:40 by shurtado          #+#    #+#             */
-/*   Updated: 2026/06/11 15:07:41 by shurtado         ###   ########.fr       */
+/*   Created: 2026/06/11 22:04:01 by shurtado          #+#    #+#             */
+/*   Updated: 2026/06/11 22:04:02 by shurtado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ bool SpectrumAnalyzer::processMessage(GstMessage* msg) {
   const guint count = gst_value_list_get_size(val);
   for (guint i = 0; i < count && i < bands_; ++i) {
     const GValue* item = gst_value_list_get_value(val, i);
-    magnitudes_[i] = static_cast<float>(g_value_get_float(item));
+    magnitudes_[i] = g_value_get_float(item);
   }
 
   return true;
